+++ addon/components/navigation-api/search-bar.js
@@ -1,0 +1,12 @@
+import Component from 'ember-component';
+import hbs from 'htmlbars-inline-precompile';
+
+/**
+ * Not operational sidebar search component
+ * @class NavigationAPI.SearchBar
+ * @constructor
+ * @extends Ember.Component
+ */
+export default Component.extend({
+  layout: hbs``
+});